+++ tests/tests_fabric/strategies/test_fsdp_integration.py
@@ -23,7 +23,11 @@ from torch.nn import Parameter
 from lightning.fabric import Fabric
 from lightning.fabric.plugins import FSDPPrecision
 from lightning.fabric.strategies import FSDPStrategy
-from lightning.fabric.utilities.imports import _TORCH_GREATER_EQUAL_1_12, _TORCH_GREATER_EQUAL_2_0
+from lightning.fabric.utilities.imports import (
+    _TORCH_GREATER_EQUAL_1_12,
+    _TORCH_GREATER_EQUAL_2_0,
+    _TORCH_GREATER_EQUAL_2_1,
+)
 from lightning.fabric.wrappers import _FabricOptimizer
 from tests_fabric.helpers.models import BoringFabric
 from tests_fabric.helpers.runif import RunIf
@@ -405,7 +409,14 @@ def test_fsdp_save_filter(tmp_path):
 @RunIf(min_torch="1.13", min_cuda_gpus=1)
 def test_fsdp_manual_activation_checkpointing():
     model = torch.nn.Sequential(torch.nn.Linear(1, 1), torch.nn.Linear(1, 1))
-    strategy = FSDPStrategy(activation_checkpointing=torch.nn.Linear)
+
+    if _TORCH_GREATER_EQUAL_2_1:
+        from torch.distributed.fsdp.wrap import ModuleWrapPolicy
+
+        strategy = FSDPStrategy(activation_checkpointing_policy=ModuleWrapPolicy({torch.nn.Linear}))
+    else:
+        strategy = FSDPStrategy(activation_checkpointing=torch.nn.Linear)
+
     fabric = Fabric(devices=1, accelerator="cuda", strategy=strategy)
     fabric.launch()
 
@@ -421,7 +432,7 @@ def test_fsdp_manual_activation_checkpointing():
     assert wrappers == {"0", "1"}
 
     # let fabric set up the model, it shouldn't apply activation checkpointing again
-    with pytest.warns(match="Linear'] is configured, but the model already contains checkpointed"):
+    with pytest.warns(match="is configured, but the model already contains checkpointed"):
         model = fabric.setup(model)
 
     wrappers = {name for name, mod in model._forward_module.named_modules() if isinstance(mod, CheckpointWrapper)}